config singleton bugfix
